allow for import of rds instances and proxies
