+++ pkg/lang/javascript/aws_runtime/clients.js
@@ -1,8 +1,8 @@
-const {S3Client} = require('@aws-sdk/client-s3')
-const {LambdaClient}  = require('@aws-sdk/client-lambda')
-const {SecretsManagerClient}  = require('@aws-sdk/client-secrets-manager')
-const { SNSClient }  = require('@aws-sdk/client-sns')
-const {DynamoDBClient}  = require('@aws-sdk/client-dynamodb')
+const { S3Client } = require('@aws-sdk/client-s3')
+const { LambdaClient } = require('@aws-sdk/client-lambda')
+const { SecretsManagerClient } = require('@aws-sdk/client-secrets-manager')
+const { SNSClient } = require('@aws-sdk/client-sns')
+const { DynamoDBClient } = require('@aws-sdk/client-dynamodb')
 const AWSXRay = require('aws-xray-sdk-core')
 
 const endpoint = process.env['AWS_ENDPOINT']
